refactor: createSurvey -> survey
