+++ src/features/survey/addSurveyForm/choiceForm/MaximumSelect.tsx
@@ -1,8 +1,8 @@
 import { type Dispatch, type SetStateAction } from 'react';
 import { css, type Theme } from '@emotion/react';
 
-import Counter from '~/features/createSurvey/addSurveyForm/choiceForm/Counter';
-import { MULTI_SELECT_MAX_COUNT, MULTI_SELECT_MIN_COUNT } from '~/features/createSurvey/constants';
+import Counter from '~/features/survey/addSurveyForm/choiceForm/Counter';
+import { MULTI_SELECT_MAX_COUNT, MULTI_SELECT_MIN_COUNT } from '~/features/survey/constants';
 import { BODY_2_BOLD } from '~/styles/typo';
 
 interface Props {
