+++ src/features/survey/questionList/QuestionWithDnd.tsx
@@ -3,8 +3,8 @@ import { css } from '@emotion/react';
 import { Reorder, useDragControls } from 'framer-motion';
 
 import MenuIcon from '~/components/icons/MenuIcon';
-import Question from '~/features/createSurvey/questionList/Question';
-import { type QuestionItem } from '~/features/createSurvey/types';
+import Question from '~/features/survey/questionList/Question';
+import { type QuestionItem } from '~/features/survey/types';
 
 interface Props {
   item: QuestionItem;
